示例App整改，加入react-navigation
