+++ Examples/LargeListExamples/HeightUnequalExample.js
@@ -3,21 +3,27 @@
  * Created by Stone
  * https://github.com/bolan9999
  * Email: shanshang130@gmail.com
- * Date: 2018/7/19
+ * Date: 2018/7/18
  *
  */
 
 import React from "react";
 import { StyleSheet, Text, View } from "react-native";
-import { LargeList } from "../src";
+import { LargeList } from "../../src";
+import { CommonLottieHeader, CommonLottieFooter } from "react-native-spring-scrollview/Customize";
 
-export class HeightEqualExample extends React.Component {
-  _sectionCount = 10;
-  _rowCount = 10;
+export class HeightUnequalExample extends React.Component {
+  static navigationOptions = {
+    title: "HeightUnequalExample"
+  };
+
+  _sectionCount = 1;
+  _rowCount = 5;
+  _list: LargeList;
 
   constructor(props) {
     super(props);
-    this.state = {};
+    this.state = { select: 0 };
   }
 
   render() {
@@ -33,15 +39,25 @@ export class HeightEqualExample extends React.Component {
       <LargeList
         style={styles.container}
         data={data}
+        ref={ref => (this._list = ref)}
         heightForSection={() => 50}
         renderSection={this._renderSection}
-        heightForIndexPath={() => 50}
+        heightForIndexPath={({ section: section, row: row }) => (row % 2 ? 50 : 100)}
         renderIndexPath={this._renderIndexPath}
-        renderHeader={this._renderHeader}
-        renderFooter={this._renderFooter}
+        refreshHeader={CommonLottieHeader}
+        loadingFooter={CommonLottieFooter}
+        onRefresh={this._onRefresh}
+        onLoading={this._onLoading}
       />
     );
   }
+
+  _onRefresh = () => {
+    setTimeout(() => this._list.endRefresh(), 2000);
+  };
+  _onLoading = () => {
+    setTimeout(() => this._list.endLoading(), 2000);
+  };
 
   _renderSection = (section: number) => {
     return (
@@ -63,31 +79,11 @@ export class HeightEqualExample extends React.Component {
       </View>
     );
   };
-
-  _renderHeader = () => {
-    return (
-      <View>
-        <Text style={styles.header}>I am header</Text>
-      </View>
-    );
-  };
-
-  _renderFooter = () => {
-    return (
-      <View>
-        <Text style={styles.header}>I am Footer</Text>
-      </View>
-    );
-  };
 }
 
 const styles = StyleSheet.create({
   container: {
     flex: 1
-  },
-  header: {
-    alignSelf:"center",
-    marginVertical: 50
   },
   section: {
     flex: 1,
@@ -108,4 +104,3 @@ const styles = StyleSheet.create({
     backgroundColor: "#EEE"
   }
 });
-
