+++ Examples/LargeListExamples/icons/icon1@2x.png
