+++ Examples/LargeListExamples/icons/icon1@3x.png
