+++ Examples/LargeListExamples/icons/icon2@2x.png
