+++ Examples/LargeListExamples/icons/icon2@3x.png
