+++ Examples/LargeListExamples/icons/icon3@2x.png
