+++ Examples/LargeListExamples/icons/icon3@3x.png
