+++ Examples/LargeListExamples/icons/icon4@2x.png
