+++ Examples/LargeListExamples/icons/icon4@3x.png
