+++ Examples/LargeListExamples/icons/icon6@2x.png
