+++ Examples/LargeListExamples/icons/icon6@3x.png
