+++ Examples/LargeListExamples/icons/icon9@3x.png
