+++ Examples/index.js
@@ -7,15 +7,7 @@
  *
  */
 
-export * from "./HeightUnequalExample";
-export * from "./HeightEqualExample";
-export * from "./MessageExample";
-export * from "./ContactExample";
-export * from "./MenuListExample";
-export * from "./RefreshAndLoadingExample";
-export * from "./IntensiveSectionExample";
-export * from "./ChatExample";
-export * from "./FlatListExample";
-export * from "./WaterfallListExample";
-export * from "./PictureExample";
-export * from "./StickyFormExample";
+export * from "./LargeListExamples";
+export * from "./StickyFormExamples";
+export * from "./WaterfallListExamples";
+export * from "./Home";
