+++ android/app/src/main/java/com/largelistdemo/MainApplication.java
@@ -3,6 +3,7 @@ package com.largelistdemo;
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
+import com.swmansion.gesturehandler.react.RNGestureHandlerPackage;
 import com.airbnb.android.react.lottie.LottiePackage;
 import com.bolan9999.SpringScrollViewPackage;
 import com.facebook.react.ReactNativeHost;
@@ -25,6 +26,7 @@ public class MainApplication extends Application implements ReactApplication {
         protected List<ReactPackage> getPackages() {
             return Arrays.<ReactPackage>asList(
                     new MainReactPackage(),
+            new RNGestureHandlerPackage(),
             new LottiePackage(),
                     new SpringScrollViewPackage()
             );
