+++ docs/en/_sidebar.md
@@ -16,4 +16,8 @@
         * [Overview](en/V3/WaterfallList/Overview)
         * [Usage](en/V3/WaterfallList/Usage)
 
+    * StickyForm
+        * [Overview](en/V3/StickyForm/Overview)
+        * [Usage](en/V3/StickyForm/Usage)
+
     * Precautions
